+++ client/src/components/pages/Authentication/Authentication.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useHistory } from 'react-router-dom';
 import axios from 'axios';
-import Login from "../../Login";
+import SignUp from "../../SignUp";
 import Btn from "../../Button";
 import Navbar from "react-bootstrap/Navbar";
 import "./Authentication.css";
@@ -30,7 +30,7 @@ const Authentication = () => {
 
   return (
     <>
-      <Login onSubmit={onSubmit} handleInputChange={handleInputChange} />
+      <SignUp onSubmit={onSubmit} handleInputChange={handleInputChange} />
       <Navbar sticky="bottom" className="footer">
         <Btn title={"Already have an account?"} href="/login" />
       </Navbar>
